place the focus keyphrase input in components
