+++ packages/js/src/classic-editor/components/metabox/index.js
@@ -1,15 +1,10 @@
 /* global wpseoScriptData wpseoAdminL10n */
 
 import { Fill } from "@wordpress/components";
-import { useDebounce } from "@wordpress/compose";
-import { useDispatch, useSelect } from "@wordpress/data";
-import { Fragment, useCallback } from "@wordpress/element";
 import { __ } from "@wordpress/i18n";
 import { SnippetEditor } from "@yoast/search-metadata-previews";
 import { GooglePreviewContainer } from "@yoast/seo-integration";
 import { SEO_STORE_NAME } from "@yoast/seo-store";
-import { PropTypes } from "prop-types";
-import { KeywordInput } from "../../../components/contentAnalysis/KeywordInput";
 import SeoAnalysis from "../seo-analysis";
 import ReadabilityAnalysis from "../readability-analysis";
 import MetaboxCollapsible from "../../../components/MetaboxCollapsible";
@@ -21,47 +16,7 @@ import SchemaTabContainer from "../../../containers/SchemaTab";
 import SEMrushRelatedKeyphrases from "../../../containers/SEMrushRelatedKeyphrases";
 import Warning from "../../../containers/Warning";
 import { EDITOR_STORE_NAME } from "../../editor-store";
-
-import { SeoResultsContainer } from "@yoast/seo-integration";
-import { ContentAnalysis } from "@yoast/analysis-report";
-
-/**
- * Creates the focus keyphrase input component.
- *
- * @param {string} focusKeyphraseInfoLink The URL for the help link.
- *
- * @returns {JSX.Element} The focus keyphrase input.
- */
-const FocusKeyphraseInput = ( { focusKeyphraseInfoLink } ) => {
-	const focusKeyphrase = useSelect( select => select( SEO_STORE_NAME ).selectKeyphrase() );
-	const displayNoKeyphraseMessage = useSelect( select => select( EDITOR_STORE_NAME ).getSEMrushNoKeyphraseMessage() );
-	const displayNoKeyphraseForTrackingMessage = useSelect( select => select( EDITOR_STORE_NAME ).hasWincherNoKeyphrase() );
-	const isSEMrushIntegrationActive = useSelect( select => select( EDITOR_STORE_NAME ).getIsSEMrushIntegrationActive() );
-	const { updateKeyphrase } = useDispatch( SEO_STORE_NAME );
-	const { setMarkerPauseStatus } = useDispatch( EDITOR_STORE_NAME );
-
-	const handleFocusKeyphraseChange = useCallback( keyphrase => updateKeyphrase( { keyphrase } ), [ updateKeyphrase ] );
-
-	const pauseMarker = useCallback( () => setMarkerPauseStatus( true ), [ setMarkerPauseStatus ] );
-	const startMarker = useCallback( () => setMarkerPauseStatus( false ), [ setMarkerPauseStatus ] );
-
-	return (
-		<KeywordInput
-			keyword={ focusKeyphrase }
-			displayNoKeyphraseMessage={ displayNoKeyphraseMessage }
-			displayNoKeyphraseForTrackingMessage={ displayNoKeyphraseForTrackingMessage }
-			isSEMrushIntegrationActive={ isSEMrushIntegrationActive }
-			helpLink={ focusKeyphraseInfoLink }
-			onFocusKeywordChange={ useDebounce( handleFocusKeyphraseChange ) }
-			onFocusKeyword={ pauseMarker }
-			onBlurKeyword={ startMarker }
-		/>
-	);
-};
-
-FocusKeyphraseInput.propTypes = {
-	focusKeyphraseInfoLink: PropTypes.string.isRequired,
-};
+import FocusKeyphraseInput from "../focus-keyphrase-input";
 
 /**
  * Creates the Metabox component.
